+++ toolkit/cli/smart-contracts-commands/src/register.rs
@@ -62,7 +62,7 @@ impl RegisterCmd {
 			&candidate_registration,
 			&payment_key,
 			&client,
-			FixedDelayRetries::two_minutes(),
+			FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 
@@ -94,7 +94,7 @@ impl DeregisterCmd {
 			&payment_signing_key,
 			self.spo_public_key,
 			&client,
-			FixedDelayRetries::two_minutes(),
+			FixedDelayRetries::five_minutes(),
 		)
 		.await?;