+++ toolkit/cli/smart-contracts-commands/src/reserve.rs
@@ -61,7 +61,7 @@ impl InitReserveCmd {
 			self.genesis_utxo,
 			&payment_key,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())
@@ -101,7 +101,7 @@ impl CreateReserveCmd {
 			self.genesis_utxo,
 			&payment_key,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())
@@ -134,7 +134,7 @@ impl DepositReserveCmd {
 			self.genesis_utxo,
 			&payment_key,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())
@@ -164,7 +164,7 @@ impl UpdateReserveSettingsCmd {
 			&payment_key,
 			self.total_accrued_function_script_hash,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())
@@ -190,7 +190,7 @@ impl HandoverReserveCmd {
 			self.genesis_utxo,
 			&payment_key,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())
@@ -224,7 +224,7 @@ impl ReleaseReserveCmd {
 			self.reference_utxo,
 			&payment_key,
 			&ogmios_client,
-			&FixedDelayRetries::two_minutes(),
+			&FixedDelayRetries::five_minutes(),
 		)
 		.await?;
 		Ok(())