+++ toolkit/offchain/src/init_governance/mod.rs
@@ -47,7 +47,7 @@ where
 			payment_key,
 			Some(genesis_utxo_id),
 			self,
-			FixedDelayRetries::two_minutes(),
+			FixedDelayRetries::five_minutes(),
 		)
 		.await
 		.map(|(_, tx)| tx)