+++ toolkit/offchain/src/register.rs
@@ -49,7 +49,7 @@ where
 			candidate_registration,
 			payment_signing_key,
 			self,
-			FixedDelayRetries::two_minutes(),
+			FixedDelayRetries::five_minutes(),
 		)
 		.await
 		.map_err(|e| OffchainError::InternalError(e.to_string()))
@@ -144,7 +144,7 @@ where
 			payment_signing_key,
 			stake_ownership_pub_key,
 			self,
-			FixedDelayRetries::two_minutes(),
+			FixedDelayRetries::five_minutes(),
 		)
 		.await
 		.map_err(|e| OffchainError::InternalError(e.to_string()))